api/requirements.txt
uvicorn==0.23.2
django==4.2.5
channels==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
mysqlclient==2.2.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.3.0

